\name{gisviz-package}
\alias{gisviz-package}
\alias{gisviz}
\docType{package}
\title{
GIS visualisation tools
}
\description{
General visualisations.
}
\details{
\tabular{ll}{
Package: \tab gisviz\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-18\cr
License: \tab GPL2\cr
}
}
\author{
ivanhanigan

Maintainer: Who to complain to <ivan.hanigan@gmail.com>

}
\references{

}
\keyword{ package }
\seealso{
}
\examples{
}
